Reviewed-on: https://codeberg.org/RuoQi/RuoQi-v/pulls/196

+++ backend/service/sys_admin_api/api/update_api_logic.v
@@ -75,10 +75,15 @@ fn update_api(mut ctx Context, req UpdateApiReq) !UpdateApiResp {
 		.set('api_group = ?', req.api_group)!
 		.set('service_name = ?', req.service_name)!
 		.set('method = ?', req.method)!
-		.set('is_required = ?', if req.is_required { 1 } else { 0 })! // true: 1 false: 0
+		.set('is_required = ?', u8(if req.is_required { 1 } else { 0 }))! // true: 1 false: 0
 		.set('updated_at = ?', time.now().format_ss())!
 		.where('id = ?', req.id)!
 		.update()!
+
+	// sql db {
+	// 	update SysApi set path = req.path, description = req.description, api_group = req.api_group,
+	// 	service_name = req.service_name, method = req.method, is_required = 1 where id == req.id
+	// }!
 
 	return UpdateApiResp{
 		msg: '更新成功'
